staticpod: move revision status configmap update to prune controller
